Fix #202.
This might have been broken for some time. By the looks of it since we
did some re-factoring to ease out the server side jQuery library and
replace it with the more lightweight cheerio library. It is more
lightweight but does not include all the features that jQuery includes.

This patch removes jQuery specific API calls and replaces them with the
equivalent cheerio ones where needed but in some places there was
redundant code that was failing so removing it entirely was the better
option.