Fix HTTP header typo
The tests were failing because we were trying to send the HTTP header 'Content-Type:' to the remote server and the requests library we are using (aptly named `request`) validates those for us. The HTTP header 'Content-Type:' does not exist but the header 'Content-Type' does! Note the `:` in the one that fails. :)