Tabs to spaces and set space width to 2
Spaces are better than tabs especially when doing open source development since a developers environment can set a tab to any number of spaces wheras a space is always just one character. I also set the tab width to 2 instead of 4 in some spaces since I like that better for javascript, there's a lot of indendation and 4 spaces seems like a waste of screen relastate to me. We should probably add a linting rule to apis.is to check for that.