Added worldfengur scraper
* added worldfengur scraper
scrapes the following:
http://www.worldfengur.com/freezone_horse.jsp?c=EN
implementation should be async as far as I understand it, I have some but limited experience with async javascript
there is one hack in this to encode utf8 to ISO 8859-1, the page expects ISO 8859-1 coded text, pls unhack if you know better way
regarding the page we scrape, for some searches we can get multiple horses, when that happens we get less info than we would if we'd lookup individual horse, because of this I wrote an asyncLoop (according to stackoverflow) which scrapes each horse individually from the multi horse result
simply returns empty if no horse was found

* rather return null than '' when params seem to be missing in the database

* added tests for worldfengur scraper

* raise timeout threshold

* accidental invalid header

* Update redis to version 2.5.3 🚀

http://greenkeeper.io/

* chore(package): update file to version 0.2.2

http://greenkeeper.io/

* chore(package): update babel-core to version 6.7.4

http://greenkeeper.io/

* Fix HTTP header typo

The tests were failing because we were trying to send the HTTP header 'Content-Type:' to the remote server and the requests library we are using (aptly named `request`) validates those for us. The HTTP header 'Content-Type:' does not exist but the header 'Content-Type' does! Note the `:` in the one that fails. :)

* Increase timeout for horses endpoint tests.

The remote server "worldfengur.com" doesn't seem to be very snappy so with complex queries it might hang for a bit before returning with the data we want. This increased timeout will make sure we don't write the test off as failed when it's just the remote server dragging its feet :)

* Tabs to spaces and set space width to 2

Spaces are better than tabs especially when doing open source development since a developers environment can set a tab to any number of spaces wheras a space is always just one character. I also set the tab width to 2 instead of 4 in some spaces since I like that better for javascript, there's a lot of indendation and 4 spaces seems like a waste of screen relastate to me. We should probably add a linting rule to apis.is to check for that.

* Minor code cleanup

* Remove console.log

* added set timeout to the last test

* refactored parse_data function
removed dirty temp variable which was making the code harder to read
joined params and labels together into list of objects and added param name to that mix
added comments here and there to explain thought process to help others visiting the code
readded asyncLoop