Use the provided element in a `.each()` call.
Something broke in either a greenkeeper pr or by some code change where
using `$(this)` in a cheerio `.each(function() {...})` context doesn't seem
to work any more.

Reason why I think it's a greenkeeper PR is because nothing that would
cause this seems to have been merged recently.

The solution is to take in the current element as a parameter and using
it instead like so: `.each(function(index, element) {...})`.

- [Cheerio `.each(...)`](https://github.com/cheeriojs/cheerio#each-functionindex-element-)

Fixes #286