Change file structure a bit
I moved the graphql-schemas to the corresponding endpoint folder.
I feel like this would influence others that open the project to add a graphql definition for their endpoints.
If we keep this in the graphql folder it might be hidden to newcomers and the graphql endpoint could get out of sync with the rest.