chore(tests): use glob instead of walking files
This patch removes the custom setup of running the files in the endpoints
folders and instead uses the unix style globbing to find the tests files.

This enables us to use builtin functions such as `describe.only` and
`it.only` to only run a single test or suite at a time as well as enables
us to maybe run tests in parallel.